+++ src/main/java/org/com/chimerdbv31/chimerkb/controller/ChimerKbController.java
@@ -1,10 +1,13 @@
 package org.com.chimerdbv31.chimerkb.controller;
 
+import java.util.List;
 import java.util.Locale;
 import javax.annotation.Resource;
 import javax.enterprise.inject.Model;
 import javax.servlet.http.HttpServletRequest;
 import org.com.chimerdbv31.chimerkb.services.ChimerKbService;
+import org.com.chimerdbv31.chimerkb.vo.ChimerKBVo;
+import org.com.chimerdbv31.common.vo.ParamVo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
@@ -18,9 +21,10 @@ public class ChimerKbController {
 	
 	private static final Logger logger = LoggerFactory.getLogger(ChimerKbController.class);
 	
-//	@Resource(name = "chimerKbService")
-//	private ChimerKbService chimerKbService;
-//        
+	@Resource(name = "chimerKbService")
+	private ChimerKbService chimerKbService;
+	
+	private ParamVo sParam;
 
 	@RequestMapping(value="/mchimerpub", method = RequestMethod.GET)
 	public ModelAndView chimerseqPage() throws RuntimeException{
@@ -28,6 +32,342 @@ public class ChimerKbController {
 
 		return result;
 	}
+
+	@RequestMapping(value="/msrstofchimerkb",method=RequestMethod.POST)
+    public ModelAndView rstChimerKB( HttpServletRequest request ) throws RuntimeException{
+        ModelAndView result = new ModelAndView("msRstOfChimerKBp");
+        
+        sParam = new ParamVo(true);
+        String activatedTab = request.getParameter("key_activated_tab");
+        String searchType = request.getParameter("key_a_search_type");
+        String dataForsearchType = request.getParameter("key_data_for_search_type");
+        
+        result.addObject("activated_tab", activatedTab);
+        result.addObject("search_type", searchType);
+        
+        sParam.setActivatedTab(activatedTab);
+        sParam.setSearchType(searchType);
+        sParam.setDataForSearchType(dataForsearchType);
+        String data = "";
+        StringBuffer queryStr = null;
+        String[] dataArr = null;
+        
+        
+        
+        switch( searchType ){
+                    case "all_genes":{
+                        
+                    };break;
+                    case "by_gene":{
+                        
+                        data = dataForsearchType;
+                        
+                        if(data.indexOf(",") > -1){
+                            dataArr = data.split(",");
+                            for(int i = 0; i < dataArr.length; i++){
+                                if(i==0){
+                                    sParam.setDataForSearchType( dataArr[i] );
+                                }else{
+                                    if(dataArr[i].equals("5")){
+                                        sParam.setGene5(true);
+                                    }
+                                    if( dataArr[i].equals("3") ){
+                                        sParam.setGene3(true);
+                                    }
+                                }
+                            }
+                        }
+                        
+                    };break;
+                    case "by_gene_pair":{
+                    };break;
+                    case "by_chr_locus":{};break;
+                }
+                data = "";
+                data = request.getParameter("key_selt_the_websource");
+                dataArr = null;
+                queryStr = new StringBuffer();
+                if(data.indexOf("all") > -1){                    
+                    
+                }else{
+                    dataArr = data.split(",");
+                    if(dataArr != null && dataArr.length > 0){
+                        for(int i = 0; i < dataArr.length; i++){
+                            switch( dataArr[i] ){
+                                case "litratr":{
+                                    queryStr.append(" 'Literature_Curation' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                                case "cosmic":{
+                                    queryStr.append(" 'Cosmic' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                                case "mrna":{
+                                    queryStr.append(" 'mRNA_Sequence' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                                case "etc":{
+                                    queryStr.append(" 'Mitelman,OMIM,GenBank' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                            }
+                        }
+                        
+                        if(queryStr.length() > 0){
+                            sParam.setQueryForWebSource(queryStr.toString());
+                        }
+                    }
+                }
+                
+                
+                data = "";
+                data = request.getParameter("key_kb_selt_the_breakpoint");
+                dataArr = null;
+                queryStr = new StringBuffer();
+                
+                if(data.indexOf("all") > -1){                    
+                    //sParam.setQueryForBreakPointType();
+                }else{
+                    dataArr = data.split(",");
+                    if(dataArr != null && dataArr.length > 0){
+                        for(int i = 0; i < dataArr.length; i++){
+                            switch( dataArr[i] ){
+                                case "genomic":{
+                                    queryStr.append(" 'Genomic' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                                case "exon":{
+                                    queryStr.append(" 'Exonic' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                            }
+                        }
+                        
+                        if(queryStr.length() > 0){
+                            sParam.setQueryForBreakPointType(queryStr.toString());
+                        }
+                        
+                    }
+                }
+                
+                
+                data = "";
+                data = request.getParameter("key_kb_selt_the_validtn_mtd");
+                dataArr = null;
+                queryStr = new StringBuffer();
+                
+                if(data.indexOf("all") > -1){                    
+                    //sParam.setQueryForValidationMtd();
+                }else{
+                    dataArr = data.split(",");
+                    if(dataArr != null && dataArr.length > 0){
+                        
+                        for(int i = 0; i < dataArr.length; i++){
+                            switch( dataArr[i] ){                                    
+                                case "sanger":{
+                                    queryStr.append(" 'Sanger sequencing' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                                case "fish":{
+                                    queryStr.append(" 'RT-PCR, Sanger sequencing' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                                case "rtpcr":{
+                                    queryStr.append(" 'RT-PCR' ");
+                                    if( i < (dataArr.length -1) ){
+                                        queryStr.append(",");
+                                    }
+                                };break;
+                                case "none":{
+                                };break;
+                            }
+                        }
+                        if(queryStr.length() > 0){
+                            sParam.setQueryForValidationMtd(queryStr.toString());
+                        }
+                    }
+                }
+                
+                
+                
+                data = "";
+                data = request.getParameter("key_flt_by_func");
+                dataArr = null;
+                queryStr = new StringBuffer();
+                if( data.indexOf("none") > -1 ){
+                    //sParam.setQueryForFilterByFunc();
+                }else{
+                    dataArr = data.split(",");
+                    if(dataArr != null && dataArr.length > 0){
+                            for(int i = 0; i < dataArr.length; i++){
+                                if(i > 0){
+                                    switch(dataArr[i]){
+                                        case "kinase":{
+                                            queryStr.append(" Kinase != 0 ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append("or");
+                                            }
+                                        };break;
+                                        case "onco":{
+                                            queryStr.append(" Oncogene != 0 ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append("or");
+                                            }
+                                        };break;
+                                        case "tumor":{
+                                            queryStr.append(" Tumor_suppressor != 0 ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append("or");
+                                            }
+                                        };break;
+                                        case "recpt":{
+                                            queryStr.append(" Receptor != 0 ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append("or");
+                                            }
+                                        };break;
+                                        case "transcript":{
+                                            queryStr.append(" Transcription_Factor != 0 ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append("or");
+                                            }
+                                        };break;
+                                    }
+                                }
+                            }
+                        if(queryStr.length() > 0){
+                            sParam.setQueryForFilterByFunc(queryStr.toString());
+                        }
+                    }
+                }
+                
+                data = "";
+                data = request.getParameter("key_flt_by_fusn_type");
+                dataArr = null;
+                queryStr = new StringBuffer();
+                if( data.indexOf("all") > -1 ){
+                    //sParam.setQueryForFusType();
+                }else{
+                    dataArr = data.split(",");
+                    if(dataArr != null && dataArr.length > 0){
+                            for(int i = 0; i < dataArr.length; i++){
+                                if(i > 0){
+                                    switch(dataArr[i]){
+                                        case "inter_chr":{
+                                            queryStr.append(" 'Inter-chr' ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append(",");
+                                            }
+                                        };break;
+                                        case "intra_chr":{
+                                            queryStr.append(" 'Intra-chr' ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append(",");
+                                            }
+                                        };break;
+                                    }
+                                }
+                            }
+                        if(queryStr.length() > 0){
+                            sParam.setQueryForFusType(queryStr.toString());
+                        }
+                    }
+                }
+                
+                data = "";
+                data = request.getParameter("key_flt_by_supted_info");
+                dataArr = null;
+                queryStr = new StringBuffer();
+                
+                if( data.indexOf("none") > -1 ){
+                    //sParam.setQueryForSupInfo();
+                }else{
+                    dataArr = data.split(",");
+                    if(dataArr != null && dataArr.length > 0){
+                            for(int i = 0; i < dataArr.length; i++){
+                                if(i > 0){
+                                    switch(dataArr[i]){
+                                        case "chimrSeq":{
+                                            queryStr.append(" ChimerSeq != '0' ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append("or");
+                                            }
+                                        };break;
+                                        case "chimrPub":{
+                                            queryStr.append(" ChimerPub != '0' ");
+                                            if( i < (dataArr.length -1) ){
+                                                queryStr.append("or");
+                                            }
+                                        };break;
+                                    }
+                                }
+                            }
+                        if(queryStr.length() > 0){
+                            sParam.setQueryForSupInfo(queryStr.toString());
+                        }
+                    }
+                }
+                List<ChimerKBVo> chimerKbLst = this.chimerKbService.getChimerKBResult(sParam);
+                result.addObject("chimerKb_lst", chimerKbLst);
+                // out query ////////////////////////////////////////////////////////////////////////////////////////////
+                String outPutQueryStr = "select distinct * from ChimerDB3.ChimerKB_ver5 where 1=1 ";
+                switch( sParam.getSearchType() ){
+                    case "by_gene":{
+                        if( sParam.isGene5() && sParam.isGene3() ){
+                            outPutQueryStr += " and H_gene = '"+sParam.getDataForSearchType()+"' or T_gene = '"+sParam.getDataForSearchType()+"' ";
+                        }else{
+                            if( sParam.isGene5() ){
+                                outPutQueryStr += " and H_gene = '"+sParam.getDataForSearchType()+"' ";
+                            }
+                            if( sParam.isGene3() ){
+                                outPutQueryStr += " and T_gene = '"+sParam.getDataForSearchType()+"' ";
+                            }
+                        }
+                    };break;
+                    case "by_gene_pair":{
+                        outPutQueryStr += " and Fusion_pair = '"+sParam.getDataForSearchType()+"' ";
+                    };break;
+                    case "by_chr_locus":{};break;
+                }
+                if( sParam.getQueryForWebSource() != null && sParam.getQueryForWebSource() != ""){
+                    outPutQueryStr += " and webSource in ("+sParam.getQueryForWebSource()+") ";
+                }
+                if( sParam.getQueryForBreakPointType() != null && sParam.getQueryForBreakPointType() != ""){
+                    outPutQueryStr += " and Breakpoint_Type in ("+sParam.getQueryForBreakPointType()+") ";
+                }
+                if( sParam.getQueryForValidationMtd() != null && sParam.getQueryForValidationMtd() != ""){
+                    outPutQueryStr += " and Validation in ("+sParam.getQueryForValidationMtd()+") ";
+                }
+                if( sParam.getQueryForFilterByFunc() != null && sParam.getQueryForFilterByFunc() != ""){
+                    outPutQueryStr += " and ("+sParam.getQueryForFilterByFunc()+") ";
+                }
+                if( sParam.getQueryForFusType() != null && sParam.getQueryForFusType() != ""){
+                    outPutQueryStr += " and Chr_info in ("+sParam.getQueryForFusType()+") ";
+                }
+                if( sParam.getQueryForSupInfo() != null && sParam.getQueryForSupInfo() != ""){
+                    outPutQueryStr += " and ("+sParam.getQueryForSupInfo()+"); ";
+                }
+                
+                result.addObject("output_query_str", outPutQueryStr);
+                // out query ////////////////////////////////////////////////////////////////////////////////////////////
+        return result;
+    }
 //	
 //	@RequestMapping(value = "/interpro", method = RequestMethod.GET)
 //	public String interproscan(Locale locale, Model model) {
